requirements.txt
lxml==4.9.3
packaging==23.2
pdf2image==1.16.3
Pillow==10.0.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-docx==1.0.0
typing_extensions==4.8.0
